refactor(web): atomize agent routes into forms/handlers/responses (#657)
agent.py mixed JSON-body parsing, thread/runtime orchestration, the SSE chat
generator, and Redirect/JSON/Template/Stream response construction.

- src/web/agent/forms.py — ChatRequest model + select_model helper.
- src/web/agent/handlers.py — page/thread/context/permission/chat orchestration
  including the SSE generator; returns domain DTOs and raises HTTPException for
  the existing API error contract (400/404/503 with {"detail": ...}).
- src/web/agent/responses.py — AgentTemplate/AgentJson/AgentRedirect/AgentStream
  + agent_response mapper.
- src/web/routes/agent.py — thin router.

Routes, SSE streaming semantics, the 100K-token guard, auto-rename, 503
agent-unavailable payloads (incl. Retry-After), JSON shapes, and redirects are
unchanged. Completes the medium-routes module list of #657. Part of epic #402.
